chore(higgs-tts): rename tokenizer.py to text_tokenizer.py
Disambiguate from audio_codec.py (the audio neural codec, also
historically called a "tokenizer" in the Higgs codebase). The text
prompt builder is now clearly the text-side adapter.
